% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{get_piped_call_var}
\alias{get_piped_call_var}
\alias{get_piped_call_var_wref}
\alias{formatmsg}
\title{Get Initial Variable |>'d into Call Chain}
\usage{
get_piped_call_var(sc)

get_piped_call_var_wref(sc)

formatmsg(ls, msg)
}
\arguments{
\item{sc}{the result of a \code{\link[=sys.call]{sys.call()}} within a \code{check_...} function created
by \code{\link[=checker]{checker()}}}
}
\description{
Defined to denoise \code{\link[=checker]{checker()}} function body when manually inspected
}
\details{
This converts a \code{\link[=sys.call]{sys.call()}} to its character representation
(\code{\link[=as.character]{as.character()}}), then takes the last bit of it (\code{\link[=tail]{tail()}}), then extracts
the innermost parenthetical expression (\code{\link[=gsub]{gsub()}})
}
