% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.data.table.R
\name{qprobs}
\alias{qprobs}
\title{Quantile Ranges from Spans}
\usage{
qprobs(spans = c(`50` = 0.5), mid = TRUE, extent = TRUE)
}
\arguments{
\item{spans}{numeric, on (0, 1); what proportion(s) of the data
to quantile at? if \code{!is.null(names(spans))}, those names will
appear on the result as \verb{q\\\%s(l|h)} ("l" for low, "h" for high)}

\item{mid}{logical; include median? when \code{TRUE}, if \code{spans} named
this will be named \code{qmed}}

\item{extent}{logical; include min and max values?}
}
\value{
numeric values on 0, 1
}
\description{
expands spans into quantile limits
}
