% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.data.table.R
\name{quantile.data.table}
\alias{quantile.data.table}
\title{Quantile \link{data.table} Groups}
\usage{
\method{quantile}{data.table}(
  x,
  j = .(value),
  keyby = setdiff(data.table::key(x), sampleby),
  sampleby,
  probs = qprobs(),
  ...
)
}
\arguments{
\item{x}{a \link{data.table}}

\item{j}{A "select" statement for \code{x}; the resulting columns
will be the target(s) of the quantiling. See \code{\link[data.table:data.table]{data.table::data.table()}}
for details on \code{j} argument; select will be grouped by \code{keyby}}

\item{keyby}{Optionally, the grouping definition. Specify \code{keyby} XOR
\code{sampleby}.}

\item{sampleby}{Optionally, the sample-defining column(s) (i.e., what gets
quantile'd over). Specify \code{keyby} XOR \code{sampleby}}

\item{probs}{numeric on \verb{[0, 1]}; n.b. the default is identical to
\code{\link[stats:quantile]{stats::quantile()}} default, though named (see \code{\link[=qprobs]{qprobs()}}). If \code{probs} is provided
with names, those names will be used for the resulting quantile columns.}

\item{...}{Other arguments to \code{\link[stats:quantile]{stats::quantile()}}.}
}
\value{
A \link{data.table}, key from \code{keyby} (provided or infered) and a
factor "measure" column (levels from \code{j} names), and column(s) for
requested quantile(s); if \code{probs} in \code{...} is named, those names will
be used for the column names
}
\description{
Applies \code{\link[=quantile]{quantile()}} to \link{data.table} groups,
across sample column(s), for target column(s). Returns results by
group and target(s).
}
\examples{
require(data.table)
data("mtcars")
dtcars <- mtcars |> as.data.table()
dtcars |> quantile(mpg, keyby=.(cyl, gear))
dtcars |> quantile(mpg, keyby=.(cyl, gear), probs=c(lo=0.05, md=0.5, hi=0.95))
dtcars |> quantile(.(mpg, wt), keyby=.(cyl, gear), probs=c(lo=0.05, md=0.5, hi=0.95))
}
