% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_extend.R
\name{sample_pair}
\alias{sample_pair}
\title{Sample Pairs of Distinct Indices}
\usage{
sample_pair(n, prob, verbose = FALSE)
}
\arguments{
\item{n}{positive integer: number of pairs to sample}

\item{prob}{a numeric vector OR an integer: the population weights for
sampling each index OR if length(prob) == 1, then the population size
(all members equally weighted)}

\item{verbose}{Print progress messages?}
}
\value{
A list with two elements, \code{source} and \code{sink}, each of length \code{n},
corresponding to the sampled pairs indices
}
\description{
Sample Pairs of Distinct Indices
}
\examples{
draws <- sample_pair(n = 100, prob = 10)

}
