% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_extend.R
\name{sample_until}
\alias{sample_until}
\title{Sample Event Times Until a Maximum}
\usage{
sample_until(maxval, meanval, sampler, ..., oversample = 0.1, verbose = FALSE)
}
\arguments{
\item{maxval}{a positive number: the maximum limit to simulate until}

\item{meanval}{a positive number: the mean value of a draw}

\item{sampler}{a function: the sampling distribution for draws}

\item{...}{any parameter arguments to \code{sampler}}

\item{oversample}{a number: the amount to over sample by when generating}

\item{verbose}{Print progress messages?}
}
\value{
A numeric vector of draws: the cumulative sum of draws from
\code{sampler} until the last value exceeds \code{maxval}. That sum is then trimmed to
values \verb{<= maxval}. The first value is \code{0}.
}
\description{
Draw from a sample process, summing up the draws, until a maximum exceeded
}
\examples{
draws <- sample_until(maxval = 100, meanval = 1, sampler = rexp, rate = 1)
draws

}
