% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_spaghetti.R
\docType{data}
\name{StatSpaghetti}
\alias{StatSpaghetti}
\title{GGProto object for \code{\link[=stat_spaghetti]{stat_spaghetti()}}}
\format{
An object of class \code{StatSpaghetti} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
StatSpaghetti
}
\description{
GGProto object for \code{\link[=stat_spaghetti]{stat_spaghetti()}}
}
\keyword{datasets}
