% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_spaghetti.R
\name{stat_spaghetti}
\alias{stat_spaghetti}
\title{Show Replicated Observations + Median}
\usage{
stat_spaghetti(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  max.lines = if (interactive()) 10 else 100,
  alpha.sample = 1/max.lines,
  alpha.central = 1,
  central.fun = median,
  ...
)
}
\description{
\code{stat_spaghetti()} extends observation groups by \code{sample} aesthetic and computes a
central indicator with \code{central.fun}
}
\examples{
require(data.table)

dt <- CJ(scenario = 1:3, sample = 1:100, simt = 0:50)
dt[, simf := 0.5*simt*runif(.N, 0.9, 1.1) + 10*scenario ]

ggplot(dt) + aes(simt, log10(simf), color = factor(scenario)) + stat_spaghetti(
 aes(sample = sample)
) + theme_minimal()
}
