% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandArgs.R
\name{store}
\alias{store}
\title{Generic Storage Wrapper}
\usage{
store(args, obj, ...)
}
\arguments{
\item{args}{character; where to store \code{obj}. Only
the last item, i.e. \code{tail(args, 1)}, is used, so
all of the script arguments can be tossed in here if
they end in the file to write to for output.
Method of storage inferred from file extension.}

\item{obj}{the thing to store; CAVEAT EMPTOR: no
checking to see if \code{obj} is reasonable to store in
\code{tail(args, 1)}.}

\item{...}{Other arguments to the inferred methods. See:
\code{\link[=saveRDS]{saveRDS()}} for .rds, \code{\link[=save]{save()}} for .rda, \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
for image formats, \code{\link[data.table:fwrite]{data.table::fwrite()}} for csv}
}
\description{
Provides a generic storage method.
}
